changed print and eval_stat switch statement

+++ interpreter/lex.yy.c
@@ -811,7 +811,7 @@ YY_RULE_SETUP
 case 2:
 YY_RULE_SETUP
 #line 10 "lexer.l"
-{yylval.Double = double(atoi(yytext)); return VALUE;}
+{yylval.Double = (double)atof(yytext); return VALUE;}
 	YY_BREAK
 case 3:
 YY_RULE_SETUP
